+++ revdep/checks/mlr3fselect/new/mlr3fselect.Rcheck/tests/testthat/teardown.R
@@ -1,0 +1,1 @@
+pmap(list(loggers, thresholds), function(l, threshold) l$set_threshold(threshold))