+++ revdep/checks/mlr3fselect/old/mlr3fselect.Rcheck/tests/testthat.R
@@ -1,0 +1,6 @@
+library(testthat)
+library(mlr3)
+library(mlr3fselect)
+library(checkmate)
+
+test_check("mlr3fselect")