+++ revdep/checks/mlr3tuning/new/mlr3tuning.Rcheck/tests/testthat.R
@@ -1,0 +1,8 @@
+if (requireNamespace("testthat", quietly = TRUE)) {
+  library(testthat)
+  library(checkmate)
+  library(mlr3)
+  library(mlr3tuning)
+
+  test_check("mlr3tuning")
+}