+++ revdep/checks/mlr3tuning/new/mlr3tuning.Rcheck/tests/testthat/test_TunerRandomSearch.R
@@ -1,0 +1,4 @@
+test_that("TunerRandomSearch", {
+  test_tuner("random_search")
+  test_tuner_dependencies("random_search")
+})