+++ revdep/checks/mlr3tuning/old/mlr3tuning.Rcheck/tests/testthat/test_mlr_tuners.R
@@ -1,0 +1,6 @@
+test_that("mlr_tuners", {
+  expect_dictionary(mlr_tuners)
+  for (key in mlr_tuners$keys()) {
+    expect_tuner(tnr(key))
+  }
+})