+++ tests/testthat/test_OptimInstanceSingleCrit.R
@@ -125,4 +125,3 @@ test_that("OptimInstaceSingleCrit does not work with codomain > 1", {
   expect_error(OptimInstanceSingleCrit$new(objective = OBJ_2D_2D,
     terminator = trm("none")), "Codomain > 1")
 })
-